chore: simplify dev runtime tests
