+++ tests/dev-generation.test.ts
@@ -1,283 +1,24 @@
-import type { RsbuildDevServer, Rspack } from '@rsbuild/core';
+import type { RsbuildDevServer } from '@rsbuild/core';
 import { describe, expect, it, rstest } from '@rstest/core';
-import type { ServerBuild } from 'react-router';
 import { loadReactRouterServerBuild } from '../src';
 import {
   createReactRouterDevRuntime,
   registerReactRouterDevRuntime,
   unregisterReactRouterDevRuntime,
-  type DevGraphChanges,
-  type DevGraphIdentity,
-  type ReactRouterDevManifest,
-  type ReactRouterDevRuntime,
 } from '../src/dev-generation';
-
-const noKnownChanges: DevGraphChanges = {
-  web: { known: false, files: new Set() },
-  node: { known: false, files: new Set() },
-};
-
-const identityByCompilation = new WeakMap<Rspack.Compilation, symbol>();
-
-const getCompilationIdentity = (compilation: Rspack.Compilation): symbol => {
-  const existing = identityByCompilation.get(compilation);
-  if (existing) {
-    return existing;
-  }
-  const identity = Symbol();
-  identityByCompilation.set(compilation, identity);
-  return identity;
-};
-
-const graphIdentity = (
-  webCompilation: Rspack.Compilation,
-  nodeCompilation: Rspack.Compilation,
-  nodeWebCompilation: Rspack.Compilation = webCompilation
-): DevGraphIdentity => ({
-  web: getCompilationIdentity(webCompilation),
-  node: getCompilationIdentity(nodeCompilation),
-  nodeWeb: getCompilationIdentity(nodeWebCompilation),
-});
-
-type TestServerBuild = ServerBuild & { marker: string };
-
-const createBuild = (
-  marker: string,
-  routeIds = ['routes/about', 'routes/home']
-): TestServerBuild =>
-  ({
-    entry: { module: { default: () => new Response() } },
-    routes: Object.fromEntries(
-      routeIds.map(routeId => [
-        routeId,
-        { module: { default: () => null } },
-      ])
-    ),
-    assets: { routes: {}, version: marker },
-    assetsBuildDirectory: '/app/build/client',
-    basename: '/',
-    future: {},
-    isSpaMode: false,
-    marker,
-    prerender: [],
-    publicPath: '/',
-    routeDiscovery: { mode: 'initial' },
-    ssr: true,
-  }) as unknown as TestServerBuild;
-
-const createRouteManifest = (
-  id: string,
-  css: string[],
-  imports: string[] = []
-): ReactRouterDevManifest['routes'][string] => ({
-  id,
-  module: `/${id}.js`,
-  hasAction: false,
-  hasLoader: false,
-  hasClientAction: false,
-  hasClientLoader: false,
-  hasClientMiddleware: false,
-  hasDefaultExport: true,
-  hasErrorBoundary: false,
-  imports,
-  css,
-});
-
-const createDevManifest = (
-  version: string,
-  css: {
-    entry?: string[];
-    routes?: Record<string, string[]>;
-    routeImports?: Record<string, string[]>;
-  } = {}
-): ReactRouterDevManifest => ({
-  version,
-  url: '/manifest',
-  entry: { module: '/entry.js', imports: [], css: css.entry ?? [] },
-  routes: Object.fromEntries(
-    Object.entries(css.routes ?? {}).map(([id, routeCss]) => [
-      id,
-      createRouteManifest(id, routeCss, css.routeImports?.[id]),
-    ])
-  ),
-});
-
-const createCompilation = (
-  name: 'web' | 'node',
-  dependencies: {
-    builds?: string[];
-    files?: string[];
-    contexts?: string[];
-    missing?: string[];
-  } = {}
-) =>
-  ({
-    name,
-    buildDependencies: new Set(dependencies.builds),
-    fileDependencies: new Set(dependencies.files),
-    contextDependencies: new Set(dependencies.contexts),
-    missingDependencies: new Set(dependencies.missing),
-  }) as unknown as Rspack.Compilation;
-
-const createStats = (compilation: Rspack.Compilation, hasErrors = false) =>
-  ({
-    compilation,
-    hasErrors: () => hasErrors,
-  }) as Rspack.Stats;
-
-const createGraphStats = (
-  webCompilation: Rspack.Compilation,
-  nodeCompilation: Rspack.Compilation,
-  errors: { web?: boolean; node?: boolean } = {}
-) =>
-  ({
-    stats: [
-      createStats(webCompilation, errors.web),
-      createStats(nodeCompilation, errors.node),
-    ],
-  }) as Rspack.MultiStats;
-
-const captureWeb = (
-  runtime: ReactRouterDevRuntime,
-  compilation: Rspack.Compilation,
-  marker: string,
-  css?: Parameters<typeof createDevManifest>[1]
-) => {
-  runtime.captureWeb(compilation, {
-    'static/js/app': createDevManifest(marker, css),
-  });
-};
-
-const createHarness = (
-  loadBundle: (entryName: string) => Promise<unknown> | unknown,
-  options: { onCssAssetOwnershipChanged?: () => void } = {}
-) => {
-  const errors: Error[] = [];
-  const warnings: string[] = [];
-  const loadBundleMock = rstest.fn(loadBundle);
-  const server = {
-    environments: {
-      node: { loadBundle: loadBundleMock },
-    },
-  } as unknown as RsbuildDevServer;
-  const runtime = createReactRouterDevRuntime({
-    server,
-    buildPlan: {
-      defaultEntryName: 'static/js/app',
-      entryNames: ['static/js/app'],
-    },
-    onEvaluationError: error => errors.push(error),
-    onCssAssetOwnershipChanged: options.onCssAssetOwnershipChanged,
-    onWarning: warning => warnings.push(warning),
-  });
-  return { errors, loadBundle: loadBundleMock, runtime, server, warnings };
-};
+import {
+  captureWeb,
+  createDevRuntimeHarness as createHarness,
+} from './dev-generation-fixtures';
+import {
+  createBuild,
+  createCompilation,
+  createGraphStats,
+  graphIdentity,
+  noKnownChanges,
+} from './dev-runtime-fixtures';
 
 describe('React Router development runtime', () => {
-  it('publishes a validated server build pinned to its exact web manifest', async () => {
-    const rawBuild = createBuild('raw');
-    const { runtime } = createHarness(() => rawBuild);
-    const web = createCompilation('web');
-    const node = createCompilation('node');
-    const manifest = createDevManifest('web-1');
-
-    runtime.beginAttempt();
-    runtime.captureWeb(web, { 'static/js/app': manifest });
-    const waiting = runtime.load();
-    await runtime.finishAttempt(
-      createGraphStats(web, node),
-      noKnownChanges,
-      graphIdentity(web, node)
-    );
-
-    const committed = await waiting;
-    expect(committed).not.toBe(rawBuild);
-    expect(committed.assets).toEqual(manifest);
-    expect(committed.assets).not.toBe(manifest);
-  });
-
-  it('detects retargeted route css ownership', async () => {
-    const onCssAssetOwnershipChanged = rstest.fn();
-    const { runtime } = createHarness(() => createBuild('build'), {
-      onCssAssetOwnershipChanged,
-    });
-    const firstWeb = createCompilation('web');
-    const firstNode = createCompilation('node');
-
-    runtime.beginAttempt();
-    captureWeb(runtime, firstWeb, 'about-css', {
-      routes: { 'routes/about': ['/assets/shared.css'] },
-    });
-    await runtime.finishAttempt(
-      createGraphStats(firstWeb, firstNode),
-      noKnownChanges,
-      graphIdentity(firstWeb, firstNode)
-    );
-
-    const nextWeb = createCompilation('web');
-    const nextNode = createCompilation('node');
-    runtime.beginAttempt();
-    captureWeb(runtime, nextWeb, 'home-css', {
-      routes: { 'routes/home': ['/assets/shared.css'] },
-    });
-    await runtime.finishAttempt(
-      createGraphStats(nextWeb, nextNode),
-      noKnownChanges,
-      graphIdentity(nextWeb, nextNode)
-    );
-
-    expect(onCssAssetOwnershipChanged).toHaveBeenCalledOnce();
-  });
-
-  it('notifies after a committed web manifest removes route or entry css ownership', async () => {
-    const onCssAssetOwnershipChanged = rstest.fn();
-    const { runtime } = createHarness(() => createBuild('build'), {
-      onCssAssetOwnershipChanged,
-    });
-    const firstWeb = createCompilation('web');
-    const firstNode = createCompilation('node');
-
-    runtime.beginAttempt();
-    captureWeb(runtime, firstWeb, 'with-css', {
-      entry: ['/assets/entry.css'],
-      routes: { 'routes/about': ['/assets/about.css'] },
-    });
-    await runtime.finishAttempt(
-      createGraphStats(firstWeb, firstNode),
-      noKnownChanges,
-      graphIdentity(firstWeb, firstNode)
-    );
-    expect(onCssAssetOwnershipChanged).not.toHaveBeenCalled();
-
-    const removedRouteCssWeb = createCompilation('web');
-    const secondNode = createCompilation('node');
-    runtime.beginAttempt();
-    captureWeb(runtime, removedRouteCssWeb, 'without-route-css', {
-      entry: ['/assets/entry.css'],
-    });
-    await runtime.finishAttempt(
-      createGraphStats(removedRouteCssWeb, secondNode),
-      noKnownChanges,
-      graphIdentity(removedRouteCssWeb, secondNode)
-    );
-    expect(onCssAssetOwnershipChanged).toHaveBeenCalledOnce();
-
-    const removedEntryCssWeb = createCompilation('web');
-    const thirdNode = createCompilation('node');
-    runtime.beginAttempt();
-    captureWeb(runtime, removedEntryCssWeb, 'without-entry-css');
-    await runtime.finishAttempt(
-      createGraphStats(removedEntryCssWeb, thirdNode),
-      noKnownChanges,
-      graphIdentity(removedEntryCssWeb, thirdNode)
-    );
-
-    expect(onCssAssetOwnershipChanged).toHaveBeenCalledTimes(2);
-    await expect(runtime.load()).resolves.toMatchObject({
-      assets: { version: 'without-entry-css' },
-    });
-  });
-
   it('publishes css-only removals when the route file overlaps node dependencies', async () => {
     const routePath = '/app/routes/about.tsx';
     const onCssAssetOwnershipChanged = rstest.fn();
